ref: Decouple git add and push functionality from the registry target
